EVG-3671: noop reorganize host rest route files
